convert contributables service to helper functions so we can use them elsewhere
